install exactly the binfmt binary we need

+++ docker/docker-init.sh
@@ -4,9 +4,17 @@ export DOCKER_CLI_EXPERIMENTAL=enabled
 export DOCKER_BUILDKIT=1
 # login to docker using environment variables
 echo "$DOCKER_PASS" | docker login --username="$DOCKER_USER" --password-stdin
-# I could detect only the platform we need, but this is fast.
 # shellcheck disable=SC1091
 source .env.platforms
-docker run --rm --privileged tonistiigi/binfmt:latest --install "$PLATFORMS"
+if [[ $PLATFORMS == linux/armhf ]]; then
+    PLATFORMS=arm
+elif [[ $(uname -m) == "x86_64" ]]; then
+    PLATFORMS=
+elif [[ $(uname -m) == "aarch64" ]]; then
+    PLATFORMS=aarch64
+fi
+if [[ -n ${PLATFORMS:-} ]]; then
+    docker run --rm --privileged tonistiigi/binfmt:latest --install "$PLATFORMS"
+fi
 # buildx requires creating a builder on a fresh system
 docker buildx create --use
